+++ src/php/Filesystem/FilesystemFacadeInterface.php
@@ -4,6 +4,8 @@ declare(strict_types=1);
 
 namespace Phel\Filesystem;
 
+use Gacela\Framework\Health\ModuleHealthCheckInterface;
+
 interface FilesystemFacadeInterface
 {
     public function addFile(string $file): void;
@@ -11,4 +13,6 @@ interface FilesystemFacadeInterface
     public function clearAll(): void;
 
     public function getTempDir(): string;
+
+    public function getHealthCheck(): ModuleHealthCheckInterface;
 }